+++ packages/engine/Source/Scene/ArcGisMapServerImageryProvider.js
@@ -58,7 +58,10 @@ import DeveloperError from "../Core/DeveloperError.js";
  * @property {number} [maximumLevel] The maximum tile level to request, or undefined if there is no maximum.  This parameter is ignored when accessing
  *                                        a tiled server.
  * @property {string} [token] The ArcGIS access token to use to access the service.
- */
+ * @property {Object} [mapServerData] This MapServer's metadata.  This can be supplied to prevent the imagery provider from making an extraneous
+ *                                    request when the application already has the metadata.
+ * @property {Object} [parameters=ArcGisMapServerImageryProvider.DefaultParameters] Additional parameters
+ *                    to pass to the ArcGIS server in tile requests and feature picking. */
 
 /**
  * Used to track creation details while fetching initial metadata
@@ -90,6 +93,9 @@ function ImageryProviderBuilder(options) {
   this.tileWidth = defaultValue(options.tileWidth, 256);
   this.tileHeight = defaultValue(options.tileHeight, 256);
   this.maximumLevel = options.maximumLevel;
+  this.token = options.token;
+  this.mapServerData = options.mapServerData;
+  this.parameters = options.parameters;
 }
 
 /**
@@ -326,6 +332,10 @@ function ArcGisMapServerImageryProvider(options) {
     this._tilingScheme.rectangle
   );
   this._layers = options.layers;
+  this._parameters = {
+    ...ArcGisMapServerImageryProvider.DefaultParameters,
+    ...(options.parameters ?? {})
+  };
   this._credit = options.credit;
   this._tileCredits = undefined;
 
@@ -459,7 +469,11 @@ function buildImageResource(imageryProvider, x, y, level, request) {
   let resource;
   if (imageryProvider._useTiles) {
     resource = imageryProvider._resource.getDerivedResource({
-      url: `tile/${level}/${y}/${x}`,
+      url:
+        `tile/${level}/${y}/${x}` +
+        (Object.keys(imageryProvider.parameters).length > 0
+          ? "?" + objectToQuery(imageryProvider.parameters)
+          : ""),
       request: request,
     });
   } else {
@@ -476,6 +490,7 @@ function buildImageResource(imageryProvider, x, y, level, request) {
       format: "png32",
       transparent: true,
       f: "image",
+      ...imageryProvider.parameters
     };
 
     if (
@@ -694,6 +709,19 @@ Object.defineProperties(ArcGisMapServerImageryProvider.prototype, {
       return this._layers;
     },
   },
+
+  /**
+   * Gets the additional parameters to pass to the ArcGIS server in tile requests and feature picking.
+   * @memberof ArcGisMapServerImageryProvider.prototype
+   *
+   * @type {Object}
+   * @readonly
+   */
+  parameters: {
+    get: function () {
+      return this._parameters;
+    },
+  },
 });
 
 /**
@@ -732,7 +760,13 @@ ArcGisMapServerImageryProvider.fromUrl = async function (url, options) {
   provider._resource = resource;
   const imageryProviderBuilder = new ImageryProviderBuilder(options);
   const useTiles = defaultValue(options.usePreCachedTilesIfAvailable, true);
-  if (useTiles) {
+  if (defined(options.mapServerData)) {
+    // Even if we already have the map server data, we defer processing it in case there are
+    // errors.  Clients must have a chance to subscribe to the errorEvent before we raise it.
+    Promise.resolve(options.mapServerData)
+      .then((data) => metadataSuccess(data, imageryProviderBuilder))
+      .catch((error) => metadataFailure(resource, error));
+  } else if (useTiles) {
     await requestMetadata(resource, imageryProviderBuilder);
   }
 
@@ -835,6 +869,7 @@ ArcGisMapServerImageryProvider.prototype.pickFeatures = function (
     imageDisplay: `${this._tileWidth},${this._tileHeight},96`,
     sr: sr,
     layers: layers,
+    ...this.parameters
   };
 
   const resource = this._resource.getDerivedResource({